Se crea la clase persona y se cambia a estudiante para que herede persona, así como se agregan ciertos cambios para que los datos tengan más sentido
