+++ src/main/java/org/example/estudiante/Estudiante.java
@@ -1,13 +1,11 @@
 package org.example.estudiante;
 
+import org.example.persona.Persona;
+
 import java.util.Date;
 
-public class Estudiante {
+public class Estudiante extends Persona {
 
-    private int id;
-    private String nombre;
-    private String apellido;
-    private Date fechaDeNacimiento;
     private Estado estado;
 
 
@@ -16,46 +14,12 @@ public class Estudiante {
     }
 
     public Estudiante() {
+        super();
     }
 
     public Estudiante(int id, String nombre, String apellido, Date fechaDeNacimiento, Estado estado) {
-        this.id = id;
-        this.nombre = nombre;
-        this.apellido = apellido;
-        this.fechaDeNacimiento = fechaDeNacimiento;
+        super(id, nombre, apellido, fechaDeNacimiento);
         this.estado = estado;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getNombre() {
-        return nombre;
-    }
-
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
-    }
-
-    public String getApellido() {
-        return apellido;
-    }
-
-    public void setApellido(String apellido) {
-        this.apellido = apellido;
-    }
-
-    public Date getFechaDeNacimiento() {
-        return fechaDeNacimiento;
-    }
-
-    public void setFechaDeNacimiento(Date fechaDeNacimiento) {
-        this.fechaDeNacimiento = fechaDeNacimiento;
     }
 
     public Estado getEstado() {
