+++ src/main/java/org/example/persona/Persona.java
@@ -1,0 +1,54 @@
+package org.example.persona;
+
+import java.util.Date;
+
+public class Persona {
+
+    private int id;
+    private String nombre;
+    private String apellido;
+    private Date fechaDeNacimiento;
+
+
+    public Persona() {
+    }
+
+    public Persona(int id, String nombre, String apellido, Date fechaDeNacimiento) {
+        this.id = id;
+        this.nombre = nombre;
+        this.apellido = apellido;
+        this.fechaDeNacimiento = fechaDeNacimiento;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getNombre() {
+        return nombre;
+    }
+
+    public void setNombre(String nombre) {
+        this.nombre = nombre;
+    }
+
+    public String getApellido() {
+        return apellido;
+    }
+
+    public void setApellido(String apellido) {
+        this.apellido = apellido;
+    }
+
+    public Date getFechaDeNacimiento() {
+        return fechaDeNacimiento;
+    }
+
+    public void setFechaDeNacimiento(Date fechaDeNacimiento) {
+        this.fechaDeNacimiento = fechaDeNacimiento;
+    }
+}
